requirements.txt
aioredis==1.3.1
amqp==2.6.1
aniso8601==9.0.1
asgiref==3.6.0
async-timeout==3.0.1
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
billiard==3.6.3.0
boto3==1.14.47
botocore==1.17.47
celery==4.4.7
certifi==2020.6.20
cffi==1.15.1
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
constantly==15.1.0
cryptography==3.0
daphne==2.5.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.10
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.10.1
django-debug-toolbar==2.2
django-graphql-jwt==0.3.4
django-import-export==3.0.2
django-storages==1.9.1
django-timezone-field==4.0
docutils==0.15.2
et-xmlfile==1.1.0
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
jmespath==0.10.0
kombu==4.6.11
MarkupPy==1.14
msgpack==1.0.0
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.4.0
promise==2.3
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.6.0
pyOpenSSL==19.1.0
pyotp==2.8.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==6.0
redis==3.5.3
requests==2.24.0
s3transfer==0.3.3
sentry-sdk==0.16.5
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
tablib==3.3.0
text-unidecode==1.3
Twisted==21.2.0
txaio==20.4.1
urllib3==1.25.10
vine==1.3.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.11.0
zope.interface==5.5.2
